WIP: get value from Token enum
Attempted to write a function for retreiving a value from an enum.

Enums could have values of different types e.g. usize, String,
or no value at all. It was found to be difficult to define a
function for returning the value as no one type could be used to
cover all the possible types returned by the function.

A fundamentaly new approach to expressing tokens will be attempted
in a new branch.